# internal/api/proto/v1/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "internal/api/proto/v1/service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BookAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/book": {
      "get": {
        "operationId": "BookAPI_GetBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "BookAPI"
        ]
      },
      "delete": {
        "operationId": "BookAPI_DeleteBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "BookAPI"
        ]
      },
      "post": {
        "operationId": "BookAPI_AddBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddBookRequest"
            }
          }
        ],
        "tags": [
          "BookAPI"
        ]
      },
      "put": {
        "operationId": "BookAPI_UpdateBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateBookRequest"
            }
          }
        ],
        "tags": [
          "BookAPI"
        ]
      }
    },
    "/books": {
      "get": {
        "operationId": "BookAPI_AllBooks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AllBooksResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "BookAPI"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddBookRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1AddBookResponse": {
      "type": "object",
      "properties": {
        "book": {
          "$ref": "#/definitions/v1Book"
        }
      }
    },
    "v1AllBooksResponse": {
      "type": "object",
      "properties": {
        "book1": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Book1"
          }
        }
      }
    },
    "v1Book": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        },
        "cteatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1Book1": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        },
        "cteatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1DeleteBookResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1GetBookResponse": {
      "type": "object",
      "properties": {
        "book": {
          "$ref": "#/definitions/v1Book"
        }
      }
    },
    "v1UpdateBookRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1UpdateBookResponse": {
      "type": "object",
      "properties": {
        "book": {
          "$ref": "#/definitions/v1Book"
        }
      }
    }
  }
}